Power Stages, NS's W stuns dogs, and tweaks
Sniper: starts with Q and E active abilities
1. GAIN Meepo Net (W)
2. GAIN Timberchain (R)
3. Bonus hounds (+2)

Nightstalker: starts with Q and W active abilities
1. GAIN Leap (E)
2. GAIN Echolocation (R)
3. Crippling Fear (W) becomes AOE

When Crippling/Paralyzing Fear is cast on dogs, then they are stunned
for the full duration. TODO: Change to a new retreating behavior.

Shadowstalker is in this commit but horribly broken and unworkable, so
it's already defunct.